Unifying schema converter test.
